fix value types
Temperature, humidity and pressure are float data types. The values were
stored as integeres in our database. This change is changing the
values's data type to a float.

+++ api/src/main/java/ch/tbzedu/weatherapp/Weather.java
@@ -24,11 +24,11 @@ public class Weather {
     private Timestamp createdAt;
 
     @Column(name = "temperature", nullable = false)
-    private int temperature;
+    private float temperature;
 
     @Column(name = "humidity", nullable = false)
-    private int humidity;
+    private float humidity;
 
     @Column(name = "pressure")
-    private int pressure;
+    private float pressure;
 }
